+++ src/scripts/create_multi_wiki_qa.py
@@ -42,6 +42,7 @@ LANGUAGES = [
     "pl",
     "pt-pt",
     "lv",
+    "lt",
 ]