+++ src/scripts/create_winogrande.py
@@ -32,6 +32,7 @@ LANGUAGES = [
     "fi",
     "fr",
     "it",
+    "lt",
     "lv",
     "nl",
     "no",